Split package list into separate file, add overrides
The package list is now a a `makeScope` scope, thus allowing overriding
of packages. It also puts the autogenerated package list into an
independent file so changes to `default.nix` can be made without
touching the generator script.

Important: This doesremove the packages from the top-level of raku-nix.
They're accessible from the `rakuPackages` attribute instead. We could
merge `rakuPackages` back into the top-level, but that wouldn't work
after any overrides are applied.

Note: The changed files were updated by hand in sync with the generator
script, as I cannot actually run the generator script due to CRAI not
existing anymore.